+++ src/dll/add_dll.c
@@ -1,0 +1,38 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   add_dll.c                                          :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: ttsubo <ttsubo@student.42.fr>              +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/01/04 15:20:59 by ttsubo            #+#    #+#             */
+/*   Updated: 2025/01/15 16:23:15 by ttsubo           ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "doubly_linked_list.h"
+
+/**
+ * @brief 双方向リストの末端にノードを追加します
+ * 
+ * @param self 
+ * @param value
+ * @note	TODO: 先頭に追加するように変更する
+ */
+void	add_dll(t_dll *self, int value)
+{
+	t_dll_node	*node;
+
+	node = (t_dll_node *)ft_calloc(sizeof(t_dll_node), 1);
+	node->value = value;
+	if (!self->head)
+	{
+		self->head = node;
+		self->tail = node;
+	}
+	self->head->prev = node;
+	node->next = self->head;
+	self->tail->next = node;
+	node->prev = self->tail;
+	self->tail = node;
+}
