+++ src/dll/init_dll.c
@@ -1,17 +1,22 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   doubly_linked_list.c                               :+:      :+:    :+:   */
+/*   init_dll.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: ttsubo <ttsubo@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/04 15:20:59 by ttsubo            #+#    #+#             */
-/*   Updated: 2025/01/09 21:59:25 by ttsubo           ###   ########.fr       */
+/*   Updated: 2025/01/15 16:23:15 by ttsubo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "doubly_linked_list.h"
 
+/**
+ * @brief 双方向リストを初期化して、メソッドをセットします。
+ * 
+ * @return t_dll* 
+ */
 t_dll	*init_dll(void)
 {
 	t_dll	*dll;
